chore: create on sale ordinals hook
