+++ src/hooks/index.js
@@ -3,3 +3,4 @@ export { default as useSticky } from "./use-sticky";
 export { default as useFlyoutSearch } from "./use-flyout-search";
 export { default as useScrollToTop } from "./use-scroll-to-top";
 export { default as useConnectWallet } from "./use-connect-wallet";
+export { default as useOnsaleOrdinals } from "./use-onsale-ordinals";
